Asignar memoria Calloc-Malloc

+++ more_malloc_free/2-calloc.c
@@ -9,12 +9,12 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 	if (nmemb == 0 || size == 0)
 		return (NULL);
 
-	p = (int *)malloc(nmemb * (int)size);
+	p = (int *)malloc(nmemb * size);
 
 	if (p == NULL)
 		return (NULL);
 
-	for (x = 0; x < nmemb; x++)
+	for (x = 0; x < nmemb;x++)
 	{
 		p[x] = 0;
 	}
